move LevenshteinDistance out of internal
